Align spell preset serialization helpers
